Move types to ecs lib folder
